refactor: gstr-1 generation and summarization
